Bug 1838415 - Extend RFPTarget size to uint64_t. r=tjr
Differential Revision: https://phabricator.services.mozilla.com/D180936

+++ dom/canvas/OffscreenCanvas.h
@@ -22,7 +22,7 @@ struct JSContext;
 namespace mozilla {
 class CancelableRunnable;
 class ErrorResult;
-enum class RFPTarget : unsigned;
+enum class RFPTarget : uint64_t;
 
 namespace gfx {
 class SourceSurface;
